feat: page de profil user

+++ delete-article.php
@@ -20,7 +20,10 @@
         $idArticle = $_GET['id'] ?? '';
 
         if($idArticle) {
-            $articleDAO->deleteOne($idArticle);
+            $article = $articleDAO->getOne($idArticle);
+            if($article['author'] === $currentUser['id']) {
+                $articleDAO->deleteOne($idArticle);
+            }
         }
 
         header('Location: /');
